reformatting + few changes

+++ firmware/include/hal.h
@@ -40,7 +40,7 @@ void pin_config(gpio_t port, pin_t pin, int options);
 /** USART functions */
 void arch_usart_send(usart_t port, uint8_t data);
 
-uint8_t arch_usart_recv(usart_t port);
+uint16_t arch_usart_recv(usart_t port);
 
 /** High level config functions */
 void config_isr(int port);
